# env/openapi/openapi.yml
...
components: 
  schemas:
    AddRentalItemEventDto: 
      required:
      - serialNo
      properties:
        itemName: 
          type: string
        serialNo: 
          type: string
      description: レンタル品登録用DTO
      type: object
...
    LocalDateTime: 
      description: 日時型フォーマット
      example: 20220904 09:00
      format: yyyyMMdd HH:ss
      type: string
...
    RentalItemResourceDto: 
      required:
      - id
      - serialNo
      properties:
        id: 
          format: int32
          type: integer
        itemName: 
          type: string
        serialNo: 
          type: string
      description: レンタル品DTO
      type: object
...
  responses:
...
    ServerError: 
      headers:
        Rms-Exception: 
          description: 発生例外のRmsSystemExceptionが設定される
          required: true
          schema: 
            type: string
          style: simple
      content:
        application/json: 
          schema: 
            $ref: '#/components/schemas/GenericErrorInfo'
      description: アプリケーション内部でエラーが発生した場合
...
    DataDupricate: 
      headers:
        Rms-Exception: 
          description: 発生例外のBusinessFlowExceptionが設定される
          required: true
          schema: 
            type: string
          style: simple
      content:
        application/json: 
          schema: 
            $ref: '#/components/schemas/GenericErrorInfo'
      description: 登録データが既に登録されている
...
    ParameterError: 
      headers:
        Rms-Exception: 
          description: 発生例外のConstraintViolationExceptionが設定される
          required: true
          schema: 
            type: string
          style: simple
      content:
        application/json: 
          schema: 
            $ref: '#/components/schemas/ValidationErrorInfoImpl'
      description: パラメータエラーの場合
  securitySchemes:
    RmsJwtAuth: 
      bearerFormat: JWT
      description: 認証と認可はMicroProfile JWT RBAC Securityの仕様をもとに行い認証エラーの場合は401を認可エラーの場合は403を返す
      in: header
      scheme: bearer
      type: http
info: 
  contact: 
    name: 課外活動
    url: https://extact-io.github.io/rms-website/
  title: レンタル予約システム公開API
  version: '-'
openapi: 3.0.3
paths:
...
  /api/rms/items: 
    post: 
      tags:
      - Admin
      security:
      - RmsJwtAuth: []
      description: シリアル番号が既に使われている場合は409を返す
      operationId: addRentalItem
      requestBody: 
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/AddRentalItemEventDto'
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/RentalItemResourceDto'
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '409': 
          $ref: '#/components/responses/DataDupricate'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: レンタル品を登録する
    get: 
      tags:
      - Member
      - Admin
      security:
      - RmsJwtAuth: []
      description: 登録されているすべてのレンタル品を取得する
      operationId: getAllRentalItems
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/RentalItemResourceDto'
                type: array
          description: 検索結果
      summary: レンタル品の全件を取得する
    put: 
      tags:
      - Admin
      security:
      - RmsJwtAuth: []
      description: 依頼されたレンタル品を更新する
      operationId: updateRentalItem
      requestBody: 
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/RentalItemResourceDto'
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/RentalItemResourceDto'
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: レンタル品を更新する
  /api/rms/items/rentable: 
    get: 
      tags:
      - Member
      parameters:
      - 
        description: 利用開始日時
        in: query
        name: from
        required: true
        schema: 
          allOf:
          - 
            $ref: '#/components/schemas/LocalDateTime'
          - 
            nullable: false
      - 
        description: 利用開始日時
        in: query
        name: to
        required: true
        schema: 
          allOf:
          - 
            $ref: '#/components/schemas/LocalDateTime'
          - 
            nullable: false
      security:
      - RmsJwtAuth: []
      description: 該当期間に予約可能なレンタル品を検索する
      operationId: findCanRentedItemAtTerm
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/RentalItemResourceDto'
                type: array
          description: 検索結果
        '400': 
          $ref: '#/components/responses/ParameterError'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 該当期間に予約可能なレンタル品を検索する
  /api/rms/items/{itemId}: 
    delete: 
      tags:
      - Admin
      parameters:
      - 
        description: レンタル品ID
        in: path
        name: itemId
        required: true
        schema: 
          format: int32
          type: integer
      security:
      - RmsJwtAuth: []
      description: 削除対象のレンタル品を参照する予約が存在する場合は削除は行わずエラーにする
      operationId: deleteRentalItem
      responses:
        '200': 
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '409': 
          $ref: '#/components/responses/DataRefered'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: レンタル品を削除する
  /api/rms/items/{itemId}/rentable: 
    get: 
      tags:
      - Member
      parameters:
      - 
        description: レンタル品ID
        in: path
        name: itemId
        required: true
        schema: 
          format: int32
          type: integer
      - 
        description: 利用開始日時
        in: query
        name: from
        required: true
        schema: 
          allOf:
          - 
            $ref: '#/components/schemas/LocalDateTime'
          - 
            nullable: false
      - 
        description: 利用開始日時
        in: query
        name: to
        required: true
        schema: 
          allOf:
          - 
            $ref: '#/components/schemas/LocalDateTime'
          - 
            nullable: false
      security:
      - RmsJwtAuth: []
      description: レンタル品が該当期間に予約可能かを返す
      operationId: canRentedItemAtTerm
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                type: boolean
          description: trueならレンタル可
        '400': 
          $ref: '#/components/responses/ParameterError'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: レンタル品が該当期間に予約可能かを返す
  /api/rms/reservations: 
    post: 
      tags:
      - Member
      security:
      - RmsJwtAuth: []
      description: 予約対象のレンタル品が存在しない場合は404を予定期間に別の予約が既に入っている場合は409を返す
      operationId: addReservation
      requestBody: 
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/AddReservationEventDto'
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ReservationResourceDto'
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '409': 
          $ref: '#/components/responses/DataDupricate'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: レンタル品を予約する
    get: 
      tags:
      - Admin
      security:
      - RmsJwtAuth: []
      description: 登録されているすべての予約を取得する
      operationId: getAllReservations
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/ReservationResourceDto'
                type: array
          description: 検索結果
      summary: 予約の全件を取得する
    put: 
      tags:
      - Admin
      security:
      - RmsJwtAuth: []
      description: 依頼された予約を更新する。ユーザアカウントとレンタル品のエンティティは更新時に使用していないためIDのみ設定すればよい
      operationId: updateReservation
      requestBody: 
        content:
          application/json: 
            schema: 
              $ref: '#/components/schemas/ReservationResourceDto'
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                $ref: '#/components/schemas/ReservationResourceDto'
          description: 登録成功。IDに対するユーザアカウントとレンタル品のエンティティは設定されて返される
        '400': 
          $ref: '#/components/responses/ParameterError'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 予約を更新する
  /api/rms/reservations/item/{itemId}: 
    get: 
      tags:
      - Member
      parameters:
      - 
        description: レンタル品ID
        in: path
        name: itemId
        required: true
        schema: 
          format: int32
          type: integer
      security:
      - RmsJwtAuth: []
      description: 指定されたレンタル品に対する予約を検索する
      operationId: findReservationByRentalItemId
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/ReservationResourceDto'
                type: array
          description: 検索結果
        '400': 
          $ref: '#/components/responses/ParameterError'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 指定されたレンタル品に対する予約を検索する
  /api/rms/reservations/item/{itemId}/startdate/{startDate}: 
    get: 
      tags:
      - Member
      parameters:
      - 
        description: レンタル品ID
        in: path
        name: itemId
        required: true
        schema: 
          format: int32
          type: integer
      - 
        description: 利用開始日
        in: path
        name: startDate
        required: true
        schema: 
          example: 20201230
          format: yyyyMMdd
          nullable: false
          type: string
      security:
      - RmsJwtAuth: []
      description: 指定されたレンタル品と利用開始日に一致する予約を検索する
      operationId: findReservationByRentalItemAndStartDate
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/ReservationResourceDto'
                type: array
          description: 検索結果
        '400': 
          $ref: '#/components/responses/ParameterError'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 指定されたレンタル品と利用開始日で予約を検索する
  /api/rms/reservations/own: 
    get: 
      tags:
      - Member
      security:
      - RmsJwtAuth: []
      description: ログインユーザが予約者となっている予約の一覧を取得する。このAPIは/reservations/reserver/{reserverId}のエイリアスとなっている
      operationId: getOwnReservations
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/ReservationResourceDto'
                type: array
          description: 検索結果
        '400': 
          $ref: '#/components/responses/ParameterError'
      summary: 自分の予約一覧を取得する
  /api/rms/reservations/own/{reservationId}: 
    delete: 
      tags:
      - Member
      parameters:
      - 
        description: 予約ID
        in: path
        name: reservationId
        required: true
        schema: 
          format: int32
          type: integer
      security:
      - RmsJwtAuth: []
      description: 依頼された予約IDに対する予約をキャンセルする。予約のキャンセルは予約した人しか行えない。他の人が予約キャンセルを行った場合は禁止操作としてエラーにする
      operationId: cancelReservation
      responses:
        '200': 
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '403': 
          $ref: '#/components/responses/Forbidden'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 予約をキャンセルする
  /api/rms/reservations/reserver/{reserverId}: 
    get: 
      tags:
      - Member
      parameters:
      - 
        description: ユーザID
        in: path
        name: reserverId
        required: true
        schema: 
          format: int32
          type: integer
      security:
      - RmsJwtAuth: []
      description: 指定されたユーザが予約者の予約を検索する
      operationId: findReservationByReserverId
      responses:
        '200': 
          content:
            application/json: 
              schema: 
                items: 
                  $ref: '#/components/schemas/ReservationResourceDto'
                type: array
          description: 検索結果
        '400': 
          $ref: '#/components/responses/ParameterError'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 指定されたユーザが予約者の予約を検索する
  /api/rms/reservations/{reservationId}: 
    delete: 
      tags:
      - Admin
      parameters:
      - 
        description: 予約ID
        in: path
        name: reservationId
        required: true
        schema: 
          format: int32
          type: integer
      security:
      - RmsJwtAuth: []
      description: 予約を削除する
      operationId: deleteReservation
      responses:
        '200': 
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: 予約を削除する
...
  /api/rms/users/{userAccountId}: 
    delete: 
      tags:
      - Admin
      parameters:
      - 
        description: ユーザID
        in: path
        name: userAccountId
        required: true
        schema: 
          format: int32
          type: integer
      security:
      - RmsJwtAuth: []
      description: 削除対象のユーザを参照する予約が存在する場合は削除は行わずエラーにする
      operationId: deleteUserAccount
      responses:
        '200': 
          description: 登録成功
        '400': 
          $ref: '#/components/responses/ParameterError'
        '404': 
          $ref: '#/components/responses/UnknownData'
        '409': 
          $ref: '#/components/responses/DataRefered'
        '500': 
          $ref: '#/components/responses/ServerError'
      summary: ユーザを削除する
...